tools/config-sanitizer/requirements.txt
PyYAML
pydantic
